feat: add reproducible environment with Nix flakes
This commit adds a flakes.nix file that declares a reproducible
environment. It is usable with the Nix[1] package manager (NixOS not
required) using the Flakes feature[2]. It also uses the devenv[3]
module for an easy way to declare installed packages, commands, tests,
tasks, and git hooks.

This reproducible environment provides a shell with all required tools
available for development and building the project.

[1] https://nixos.org/
[2] https://nixos.wiki/wiki/flakes
[3] https://devenv.sh/